shoppiq: Switch to spring data jpa

+++ crud/shoppiq/src/main/java/com/crud/project/shoppiq/services/ItemService.java
@@ -1,11 +1,13 @@
 package com.crud.project.shoppiq.services;
 
+import com.crud.project.shoppiq.exceptions.ItemNotFoundException;
 import com.crud.project.shoppiq.models.Item;
 import com.crud.project.shoppiq.repositories.ItemRepository;
-import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -14,23 +16,43 @@ public class ItemService {
     @Autowired
     private ItemRepository itemRepository;
 
-    @Transactional
     public Optional<Item> saveNewItem(Item newItem) {
-        return Optional.ofNullable(itemRepository.save(newItem));
+        return Optional.of(itemRepository.save(newItem));
     }
 
-    @Transactional
     public Optional<Item> getItemById(long id) {
-        return itemRepository.findById(id);
+        Optional<Item> currentItem = itemRepository.findById(id);
+        if (currentItem.isPresent()) {
+            return currentItem;
+        }
+
+        throw new ItemNotFoundException("Item with id: " + id + " not found");
     }
 
-    @Transactional
+    public Optional<List<Item>> getAllExistingItems() {
+        List<Item> itemList = new ArrayList<>();
+        itemRepository.findAll().forEach(itemList::add);
+        return Optional.of(itemList);
+    }
+
     public void deleteItemById(long id) {
-        itemRepository.deleteById(id);
+        Optional<Item> currentItem = itemRepository.findById(id);
+        if (currentItem.isPresent()) {
+            itemRepository.deleteById(id);
+            return;
+        }
+
+        throw new ItemNotFoundException("Item with id: " + id + " not found");
     }
 
-    @Transactional
     public Optional<Item> updateItemById(long id, Item newItem) {
-        return itemRepository.updateById(id, newItem);
+        Optional<Item> currentItem = itemRepository.findById(id);
+        if (currentItem.isPresent()) {
+            currentItem.get().update(newItem);
+            itemRepository.save(currentItem.get());
+            return currentItem;
+        }
+
+        throw new ItemNotFoundException("Item with id: " + id + " not found");
     }
 }
